+++ flicker/__init__.py
@@ -16,5 +16,5 @@
 import os
 
 from flicker.version import __version__
-from flicker.flicker import FlickerDataFrame
+from flicker.dataframe import FlickerDataFrame
 from flicker.group import FlickerGroupedData
